Code for Welcome Page
I am in the process of creating an interactive welcome page that asks for and stores the user's name and date of birth in the program. I think there are a lot of errors in the code (e.g. I still need to define nameBox) that I will work on debugging throughout the day. I got the welcome page to appear on the screen when I pressed "Play."