Debugging Welcome Page
I had an error in line 16, and I realized that I did not define nameBox above. I added a uicontrol function for the nameBox and ran the program again. The welcome page appeared, and when I entered my name and DOB, the program proceeded to the next page. Now, we would have to add codes that load the database onto the program before we work on the search engine. We decided to work on separate scripts first and then combine the code to minimize confusion.