Debugging Alicia's Code
I tried to fix the code that Alicia worked on as she was having trouble downloading the Kaggle dataset. With communication, we decided to try another method. I downloaded the file with all the movies and placed it in the documents section of the repository. Then, I attempted to extract the movie name, genre, and director columns of the dataset. I turned the columns of data into a single cell array so that the search system could more easily find the movie that matches the search criteria. This is a draft of the code, and I will continue to comment/edit it for better readability.