Debugging
I spent about 3 hours throughout the day trying to debug my movie ratings code because there was an error with the movies that did not have ratings. I also looked at and tried to debug Yoojin's code for the list of movies because it was only displaying war movies instead of the comprehensive list.