Separate Search System
We each tried to create a separate search system to test which code would work. I spent 3 hours coding this and testing it, but the program does not work. It defaults to "No movies found" or gives a random list of movies that are irrelevant to the search. I will try to debug this code or look at the other codes that my teammates worked on to progress on from here.