Deleting MovieRatingsVisual and Debugging the Interactive
I spent around 2 hours debugging the InteractiveWithUser platform because it kept repeating on a loop. I discovered that the error was due to there being a loop in the interactive movie search, so I made some edits to Faith's code for that so it wouldn't repeat forever. I also deleted the movie ratings visual because it was not working and it was not part of our project proposal anyway.