Updating the Search to Integrate Fuzzy Search
I added codes so that even if the input for genre or the director is not correctly spaced, the approximate results will show using the formula for Levenshtein distance. This took me about 4 hours. I still need to edit the code so that even if the input is misspelled, the approximate outputs will show.